♻️(frontend) refactor video filters logic
Move the filtering logic to the video part of the ui package,
to prepare its future refactoring in a dedicated package.

Plus introduce a new feature, a refresh button, that help
refreshing data of the current view.

Finally, add styling on the Filters component.
