+++ src/frontend/packages/ui/video/components/Filters/_index.scss
@@ -1,0 +1,22 @@
+
+
+.c__filters {
+  display: flex;
+  gap: 0.5rem;
+  margin-bottom: 1rem;
+
+  &__range-picker {
+    // FIXME - Fix Cunningham flex layout with Date Range Picker.
+    min-width: var(--c--components--forms-field--width);
+    width: auto;
+  }
+
+  &__refresh {
+    // FIXME - Introduce a 'large' mode for button component in Cunningham.
+
+    &.c__button--medium.c__button--icon-only {
+      height: 56px;
+      width: 56px;
+    }
+  }
+}
